backend/requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.32.1

# Database & ORM
sqlmodel==0.0.22
alembic==1.14.0
psycopg2-binary==2.9.10

# Supabase client (for Storage)
supabase==2.10.0

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.20

# Settings
pydantic-settings==2.7.0

# HTTP client — pinned to satisfy supabase<0.28 constraint
httpx==0.27.2

# Caching (TTL cache for exchange rates)
cachetools==5.5.0

# Email
fastapi-mail==1.4.1

# OCR (heavy — uncomment in Phase 4)
# easyocr==1.7.2

# Dev / Testing
pytest==8.3.4
pytest-asyncio==0.25.0